+++ CHANGELOG.md
@@ -7,6 +7,11 @@ on the willingness of the author to come up with any and stick to it.
 
 ## [Unreleased]
 
+### Changed
+
+* Tabs will now be rendered using 4 spaces, instead of showing up as `^I`. This
+  will be made configurable in the future.
+
 ## [0.0.3] - 2022-05-15
 ### Added